Add intermediate class for online players
