+++ src/main/java/com/faforever/client/fx/OnlinePlayerContextMenuController.java
@@ -1,23 +1,20 @@
 package com.faforever.client.fx;
 
-
 import com.faforever.client.avatar.AvatarService;
+import com.faforever.client.chat.InitiatePrivateChatEvent;
 import com.faforever.client.domain.AvatarBean;
 import com.faforever.client.domain.GameBean;
 import com.faforever.client.domain.PlayerBean;
 import com.faforever.client.game.JoinGameHelper;
 import com.faforever.client.game.PlayerStatus;
 import com.faforever.client.i18n.I18n;
-import com.faforever.client.main.event.ShowUserReplaysEvent;
 import com.faforever.client.moderator.ModeratorService;
 import com.faforever.client.notification.NotificationService;
-import com.faforever.client.player.PlayerInfoWindowController;
 import com.faforever.client.player.PlayerService;
 import com.faforever.client.player.SocialStatus;
 import com.faforever.client.replay.ReplayService;
 import com.faforever.client.reporting.ReportDialogController;
 import com.faforever.client.theme.UiService;
-import com.faforever.client.util.ClipboardUtil;
 import com.faforever.commons.api.dto.GroupPermission;
 import com.faforever.commons.lobby.GameType;
 import com.google.common.eventbus.EventBus;
@@ -26,16 +23,7 @@ import javafx.beans.WeakInvalidationListener;
 import javafx.beans.binding.Bindings;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.event.ActionEvent;
-import javafx.scene.control.Button;
-import javafx.scene.control.ColorPicker;
-import javafx.scene.control.ComboBox;
-import javafx.scene.control.ContextMenu;
-import javafx.scene.control.CustomMenuItem;
-import javafx.scene.control.MenuItem;
-import javafx.scene.control.SeparatorMenuItem;
 import javafx.scene.image.ImageView;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
@@ -52,51 +40,26 @@ import static com.faforever.client.player.SocialStatus.SELF;
 @Slf4j
 @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
 @Component
-@RequiredArgsConstructor
-public abstract class PlayerContextMenuController implements Controller<ContextMenu> {
+public abstract class OnlinePlayerContextMenuController extends AbstractPlayerContextMenuController {
 
-  private final AvatarService avatarService;
-  protected final EventBus eventBus;
-  protected final I18n i18n;
-  private final JoinGameHelper joinGameHelper;
-  protected final ModeratorService moderatorService;
-  private final NotificationService notificationService;
-  private final PlayerService playerService;
-  private final ReplayService replayService;
-  protected final UiService uiService;
-
-  public ContextMenu playerContextMenuRoot;
-  public MenuItem showUserInfo;
-  public MenuItem sendPrivateMessageItem;
-  public MenuItem copyUsernameItem;
-  public CustomMenuItem colorPickerMenuItem;
-  public ColorPicker colorPicker;
-  public Button removeCustomColorButton;
-  public SeparatorMenuItem socialSeparator;
-  public MenuItem inviteItem;
-  public MenuItem addFriendItem;
-  public MenuItem removeFriendItem;
-  public MenuItem addFoeItem;
-  public MenuItem removeFoeItem;
-  public MenuItem reportItem;
-  public SeparatorMenuItem gameSeparator;
-  public MenuItem joinGameItem;
-  public MenuItem watchGameItem;
-  public MenuItem viewReplaysItem;
-  public SeparatorMenuItem moderatorActionSeparator;
-  public MenuItem kickGameItem;
-  public MenuItem kickLobbyItem;
-  public MenuItem broadcastMessage;
-  public CustomMenuItem avatarPickerMenuItem;
-  public ComboBox<AvatarBean> avatarComboBox;
-
-  protected PlayerBean player;
   private InvalidationListener playerPropertyInvalidationListener;
+
+  public OnlinePlayerContextMenuController(AvatarService avatarService,
+                                          EventBus eventBus,
+                                          I18n i18n,
+                                          JoinGameHelper joinGameHelper,
+                                          ModeratorService moderatorService,
+                                          NotificationService notificationService,
+                                          PlayerService playerService,
+                                          ReplayService replayService,
+                                          UiService uiService) {
+    super(avatarService, eventBus, i18n, joinGameHelper, moderatorService, notificationService, playerService, replayService, uiService);
+  }
 
   @Override
   public void initialize() {
-    showUserInfo.setOnAction(event -> onShowUserInfoSelected());
-    copyUsernameItem.setOnAction(event -> onCopyUsernameSelected());
+    super.initialize();
+    sendPrivateMessageItem.setOnAction(event -> onSendPrivateMessageSelected());
     addFriendItem.setOnAction(event -> onAddFriendSelected());
     removeFriendItem.setOnAction(event -> onRemoveFriendSelected());
     addFoeItem.setOnAction(event -> onAddFoeSelected());
@@ -104,19 +67,17 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
     reportItem.setOnAction(event -> onReport());
     joinGameItem.setOnAction(event -> onJoinGameSelected());
     watchGameItem.setOnAction(event -> onWatchGameSelected());
-    viewReplaysItem.setOnAction(event -> onViewReplaysSelected());
     kickGameItem.setOnAction(event -> onKickGame());
     kickLobbyItem.setOnAction(event -> onKickLobby());
 
     avatarComboBox.setCellFactory(param -> avatarCell());
     avatarComboBox.setButtonCell(avatarCell());
     avatarPickerMenuItem.visibleProperty().bind(Bindings.createBooleanBinding(() ->
-            !avatarComboBox.getItems().isEmpty() && player.getSocialStatus() == SELF,
+            !avatarComboBox.getItems().isEmpty() && getPlayer().getSocialStatus() == SELF,
         avatarComboBox.getItems()));
 
     initializeListener();
 
-    copyUsernameItem.setVisible(true);
     socialSeparator.visibleProperty().bind(addFriendItem.visibleProperty()
         .or(removeFriendItem.visibleProperty())
         .or(addFoeItem.visibleProperty())
@@ -141,6 +102,7 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
 
   protected void initializeListener() {
     playerPropertyInvalidationListener = observable -> {
+      PlayerBean player = getPlayer();
       setModeratorOptions(moderatorService.getPermissions(), player);
       SocialStatus socialStatus = player.getSocialStatus();
       PlayerStatus playerStatus = player.getStatus();
@@ -154,7 +116,7 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
 
   protected void setItemVisibility(SocialStatus socialStatus, PlayerStatus playerStatus, GameBean game) {
     JavaFxUtil.runLater(() -> {
-      showUserInfo.setVisible(true);
+      sendPrivateMessageItem.setVisible(socialStatus != SELF);
       addFriendItem.setVisible(socialStatus != FRIEND && socialStatus != SELF);
       removeFriendItem.setVisible(socialStatus == FRIEND);
       addFoeItem.setVisible(socialStatus != FOE && socialStatus != SELF);
@@ -164,7 +126,6 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
           && (playerStatus == PlayerStatus.LOBBYING || playerStatus == PlayerStatus.HOSTING)
           && game != null && game.getGameType() != GameType.MATCHMAKER);
       watchGameItem.setVisible(playerStatus == PlayerStatus.PLAYING);
-      viewReplaysItem.setVisible(true);
     });
   }
 
@@ -183,8 +144,8 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
             .findFirst()
             .orElse(null));
 
-        // Only after the box has been populated and we selected the current value, we add the listener.
-        // Otherwise the code above already triggers a changeAvatar()
+        // Only after the box has been populated, and we selected the current value, we add the listener.
+        // Otherwise, the code above already triggers a changeAvatar()
         avatarComboBox.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, newValue) -> {
           player.setAvatar(newValue);
           avatarService.changeAvatar(Objects.requireNonNullElse(newValue, noAvatar));
@@ -193,12 +154,9 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
     });
   }
 
-  public ContextMenu getContextMenu() {
-    return getRoot();
-  }
-
+  @Override
   public void setPlayer(PlayerBean player) {
-    this.player = player;
+    super.setPlayer(player);
 
     setModeratorOptions(moderatorService.getPermissions(), player);
 
@@ -208,19 +166,12 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
     JavaFxUtil.addAndTriggerListener(player.statusProperty(), weakPlayerPropertyListener);
   }
 
-
-  public void onShowUserInfoSelected() {
-    PlayerInfoWindowController playerInfoWindowController = uiService.loadFxml("theme/user_info_window.fxml");
-    playerInfoWindowController.setPlayer(player);
-    playerInfoWindowController.setOwnerWindow(getRoot().getOwnerWindow());
-    playerInfoWindowController.show();
-  }
-
-  public void onCopyUsernameSelected() {
-    ClipboardUtil.copyToClipboard(player.getUsername());
+  public void onSendPrivateMessageSelected() {
+    eventBus.post(new InitiatePrivateChatEvent(getPlayer().getUsername()));
   }
 
   public void onAddFriendSelected() {
+    PlayerBean player = getPlayer();
     if (player.getSocialStatus() == FOE) {
       playerService.removeFoe(player);
     }
@@ -228,10 +179,12 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
   }
 
   public void onRemoveFriendSelected() {
+    PlayerBean player = getPlayer();
     playerService.removeFriend(player);
   }
 
   public void onAddFoeSelected() {
+    PlayerBean player = getPlayer();
     if (player.getSocialStatus() == FRIEND) {
       playerService.removeFriend(player);
     }
@@ -239,21 +192,25 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
   }
 
   public void onRemoveFoeSelected() {
+    PlayerBean player = getPlayer();
     playerService.removeFoe(player);
   }
 
   public void onReport() {
     ReportDialogController reportDialogController = uiService.loadFxml("theme/reporting/report_dialog.fxml");
+    PlayerBean player = getPlayer();
     reportDialogController.setOffender(player);
     reportDialogController.setOwnerWindow(getRoot().getOwnerWindow());
     reportDialogController.show();
   }
 
   public void onJoinGameSelected() {
+    PlayerBean player = getPlayer();
     joinGameHelper.join(player.getGame());
   }
 
   public void onWatchGameSelected() {
+    PlayerBean player = getPlayer();
     try {
       replayService.runLiveReplay(player.getGame().getId());
     } catch (Exception e) {
@@ -262,16 +219,12 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
     }
   }
 
-  public void onViewReplaysSelected() {
-    eventBus.post(new ShowUserReplaysEvent(player.getId()));
-  }
-
   public void onKickGame() {
-    moderatorService.closePlayersGame(player);
+    moderatorService.closePlayersGame(getPlayer());
   }
 
   public void onKickLobby() {
-    moderatorService.closePlayersLobby(player);
+    moderatorService.closePlayersLobby(getPlayer());
   }
 
   protected void setModeratorOptions(Set<String> permissions, PlayerBean player) {
@@ -281,14 +234,5 @@ public abstract class PlayerContextMenuController implements Controller<ContextM
       kickGameItem.setVisible(notSelf & permissions.contains(GroupPermission.ADMIN_KICK_SERVER));
       kickLobbyItem.setVisible(notSelf & permissions.contains(GroupPermission.ADMIN_KICK_SERVER));
     });
-  }
-
-  @Override
-  public ContextMenu getRoot() {
-    return playerContextMenuRoot;
-  }
-
-  public void consumer(ActionEvent actionEvent) {
-    actionEvent.consume();
   }
 }
