+++ src/test/java/com/faforever/client/fx/AbstractPlayerContextMenuControllerTest.java
@@ -1,7 +1,6 @@
-package com.faforever.client.teammatchmaking;
+package com.faforever.client.fx;
 
 import com.faforever.client.avatar.AvatarService;
-import com.faforever.client.builders.AvatarBeanBuilder;
 import com.faforever.client.builders.PlayerBeanBuilder;
 import com.faforever.client.domain.GameBean;
 import com.faforever.client.domain.PlayerBean;
@@ -19,17 +18,11 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.testfx.util.WaitForAsyncUtils;
 
-import java.net.URL;
-import java.util.Arrays;
-import java.util.concurrent.CompletableFuture;
-
-import static com.faforever.client.player.SocialStatus.OTHER;
 import static com.faforever.client.player.SocialStatus.SELF;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Mockito.when;
 
-public class PartyMemberContextMenuControllerTest extends UITest {
+public class AbstractPlayerContextMenuControllerTest extends UITest {
   private static final String TEST_USER_NAME = "junit";
 
   @Mock
@@ -50,21 +43,16 @@ public class PartyMemberContextMenuControllerTest extends UITest {
   private ReplayService replayService;
   @Mock
   private UiService uiService;
-  
-  private PartyMemberContextMenuController instance;
-  private PlayerBean player;
+
+  private AbstractPlayerContextMenuController instance;
 
   @BeforeEach
   public void setUp() throws Exception {
-    when(avatarService.getAvailableAvatars()).thenReturn(CompletableFuture.completedFuture(Arrays.asList(
-        AvatarBeanBuilder.create().defaultValues().url(new URL("http://www.example.com/avatar1.png")).description("Avatar Number #1").get(),
-        AvatarBeanBuilder.create().defaultValues().url(new URL("http://www.example.com/avatar2.png")).description("Avatar Number #2").get(),
-        AvatarBeanBuilder.create().defaultValues().url(new URL("http://www.example.com/avatar3.png")).description("Avatar Number #3").get()
-    )));
-    instance = new PartyMemberContextMenuController(avatarService, eventBus, i18n, joinGameHelper, moderatorService, notificationService, playerService, replayService, uiService);
+    instance = new AbstractPlayerContextMenuController(avatarService, eventBus, i18n, joinGameHelper, moderatorService,
+        notificationService, playerService, replayService, uiService) {};
     loadFxml("theme/player_context_menu.fxml", clazz -> instance, instance);
 
-    player = PlayerBeanBuilder.create().defaultValues().username(TEST_USER_NAME).socialStatus(SELF).avatar(null).game(new GameBean()).get();
+    PlayerBean player = PlayerBeanBuilder.create().defaultValues().username(TEST_USER_NAME).socialStatus(SELF).avatar(null).game(new GameBean()).get();
     instance.setPlayer(player);
     WaitForAsyncUtils.waitForFxEvents();
   }
@@ -81,30 +69,6 @@ public class PartyMemberContextMenuControllerTest extends UITest {
     assertThat(instance.addFoeItem.isVisible(), is(false));
     assertThat(instance.removeFoeItem.isVisible(), is(false));
     assertThat(instance.reportItem.isVisible(), is(false));
-    assertThat(instance.joinGameItem.isVisible(), is(false));
-    assertThat(instance.watchGameItem.isVisible(), is(false));
-    assertThat(instance.viewReplaysItem.isVisible(), is(true));
-    assertThat(instance.kickGameItem.isVisible(), is(false));
-    assertThat(instance.kickLobbyItem.isVisible(), is(false));
-    assertThat(instance.broadcastMessage.isVisible(), is(false));
-    assertThat(instance.avatarPickerMenuItem.isVisible(), is(true));
-  }
-
-  @Test
-  public void testShowCorrectItemsForOther() {
-    player.setSocialStatus(OTHER);
-    WaitForAsyncUtils.waitForFxEvents();
-
-    assertThat(instance.showUserInfo.isVisible(), is(true));
-    assertThat(instance.sendPrivateMessageItem.isVisible(), is(true));
-    assertThat(instance.copyUsernameItem.isVisible(), is(true));
-    assertThat(instance.colorPickerMenuItem.isVisible(), is(false));
-    assertThat(instance.inviteItem.isVisible(), is(false));
-    assertThat(instance.addFriendItem.isVisible(), is(true));
-    assertThat(instance.removeFriendItem.isVisible(), is(false));
-    assertThat(instance.addFoeItem.isVisible(), is(true));
-    assertThat(instance.removeFoeItem.isVisible(), is(false));
-    assertThat(instance.reportItem.isVisible(), is(true));
     assertThat(instance.joinGameItem.isVisible(), is(false));
     assertThat(instance.watchGameItem.isVisible(), is(false));
     assertThat(instance.viewReplaysItem.isVisible(), is(true));
